relocate qa samples and fix paths
