+++ tests/Overview.test.js
@@ -20,11 +20,6 @@ import ImageGallery from '../client/src/components/Overview/components/ImageGall
 import ProductInfo from '../client/src/components/Overview/components/ProductInfo.jsx';
 import StyleSelector from '../client/src/components/Overview/components/StyleSelector.jsx';
 
-import reviewData from './sample-review-data.js';
-import Ratings from '../client/src/components/Ratings/Ratings.jsx';
-import RatingsList from '../client/src/components/Ratings/components/RatingsList.jsx';
-import IndividualReview from '../client/src/components/Ratings/components/IndividualReview.jsx';
-
 jest.mock('axios');
 // beforeAll((done) => {
 //   done();
@@ -36,7 +31,7 @@ jest.mock('axios');
 
 describe('Overview', function() {
 
-  xdescribe('Render Overview Widget', function() {
+  describe('Render Overview Widget', function() {
 
     test('should Render the Overview Component', function() {
       // axios.get.mockResolvedValueOnce(productData);
@@ -68,56 +63,5 @@ describe('Overview', function() {
       const StyleSelectorElement = screen.getByTestId('style-selector');
       expect(StyleSelectorElement).toBeInTheDocument();
     })
-  })
-})
-
-// ========== Related Products ========== //
-xdescribe('Related Products', function() {
-  describe('SUB-SECTION EXAMPLE', function() {
-
-    test('TEST EXAMPLE', function() {
-      var testsWorking = true;
-
-      expect(testsWorking).toBe(true);
-    })
-
-  })
-})
-
-// ========== Questions & Answers ========== //
-xdescribe('Questions & Answers', function() {
-  describe('SUB-SECTION EXAMPLE', function () {
-
-    test('TEST EXAMPLE', function () {
-      var testsWorking = true;
-
-      expect(testsWorking).toBe(true);
-    })
-
-  })
-})
-
-// ========== Ratings & Reviews ========== //
-xdescribe('Ratings & Reviews', function() {
-  describe('Render Ratings Component', function () {
-
-    test('should Render the Ratings Component', function() {
-      axios.get.mockResolvedValueOnce(reviewData);
-      render(<Ratings />);
-      expect(screen.getByText('Ratings & Reviews')).toBeInTheDocument();
-    })
-
-    test('should Render the Ratings List Component', function() {
-      render(<RatingsList />);
-      const RatingsListElement = screen.getByTestId('ratings-list');
-      expect(RatingsListElement).toBeInTheDocument();
-    })
-
-    test('should Render the individual Review Component', function() {
-      render(<individualReview />);
-      const individualReviewElement = screen.getByTestId('individual-review');
-      expect(individualReviewElement).toBeInTheDocument();
-    })
-
   })
 })
