+++ tests/QA.test.js
@@ -7,7 +7,7 @@ import _ from 'underscore';
 import "regenerator-runtime/runtime.js";
 import '@testing-library/jest-dom';
 import {render, screen, cleanup, fireEvent, waitFor} from '@testing-library/react';
-import {questionData} from './samples/sampleData.js';
+import {questionData} from './samples/sample-qa-data.js';
 
 import QA from '../client/src/components/QA/QA.jsx';
 import AnswerQuestion from '../client/src/components/QA/components/AnswerQuestion.jsx';
@@ -15,6 +15,15 @@ import Answers from '../client/src/components/QA/components/Answers.jsx';
 import AskQuestion from '../client/src/components/QA/components/AskQuestion.jsx';
 import Questions from '../client/src/components/QA/components/Questions.jsx';
 import SearchQuestion from '../client/src/components/QA/components/SearchQuestion.jsx';
+
+jest.mock('axios');
+// beforeAll((done) => {
+//   done();
+// })
+
+// afterAll((done) => {
+//   done();
+// })
 
 describe('Questions & Answers', function() {
